asset-manager: init: erase partition table on failure
If cURL fails the image download, the disk can be left in an
inconsistent state, with a partially written OS that's bootable, but
unusable. This can result in the client system no longer PXE booting,
and unable to boot the OS from disk.

When cURL fails, erase the partition table so that the client can still
fallback to network booting.

Signed-off-by: Joseph Kogut <joseph@balena.io>